requirements.txt
Flask==2.3.3
tensorflow-cpu==2.15.0
numpy>=1.23.5
Pillow>=10.0.0
opencv-python-headless>=4.8.0
protobuf>=3.20.0,<4.0.0
python-multipart==0.0.6
gunicorn==21.2.0 